refactor(service): converge param help onto a single fieldFacts list
The two help surfaces — the typed flag's usage line and the params-only
--params addendum — each decided independently which facts a param shows
and how, and had already drifted: the addendum rendered enums values-only
(no meanings) and silently lacked the API default.

fieldFacts is now the one list of a param's facts (description, enum,
bounds, API default): the usage line joins it inline, the addendum prints
one line per fact, and a test pins that both surfaces render every fact —
so a future fact (e.g. a deprecation marker) is added in one place and
cannot reach one surface but not the other.

The two sanitizers collapse into one inlineClause pipeline with the two
policies (enum-option: first clause, 40 runes; field description: full
sentences up to a note separator, 60 runes) as documented one-line
wrappers. Help rendering moves to paramhelp.go; paramflags.go is binder
mechanics (registration, collision, overlay) only.
